instead to iterate over logger to close them use the logging function shutdow
to avoid leak of file handle and avoid "too many files open" error
the logger hamdlers must be closed after each unit test, but instead of itaerte
over loggers and close them manually I use the the logging function
shutdown.

the loggers handlers  are also close if an option doe not validate in Config

due to loogers implementation as singleton,
the unit test of config induce reusing loggers in each tests.
some tests have side effects on loggers of the other tests.
to avoid this. the loggers are remove afetr each tests.